Centralize dashboard and analytics types
