+++ src/components/dashboard/admin/kpi/AbsenteeismKpiCard.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Users } from 'lucide-react';
-import { DashboardWidgetConfig } from '@/types/dashboard';
+import { DashboardWidgetConfig } from '@types/dashboard';
 
 interface AbsenteeismKpiCardProps {
   widget: DashboardWidgetConfig;
