+++ src/components/dashboard/admin/kpi/ProductivityKpiCard.tsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { LineChart } from 'lucide-react';
-import { DashboardWidgetConfig } from '@/types/dashboard';
+import { DashboardWidgetConfig } from '@types/dashboard';
 
 interface ProductivityKpiCardProps {
   widget: DashboardWidgetConfig;
