+++ src/contexts/OrchestrationContext.tsx
@@ -9,7 +9,7 @@ import {
   SanctuaryWidget,
   EmotionalSynthesis,
   OrchestrationEvent
-} from '@/types/global';
+} from '@types/orchestration';
 import { EmotionResult } from '@/types/emotion';
 
 // Données initiales pour le sanctuaire
