+++ tsconfig.app.json
@@ -23,7 +23,8 @@
 
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@types/*": ["./types/*"]
     }
   },
   "include": ["src", "types"]
