+++ tsconfig.json
@@ -7,7 +7,8 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@types/*": ["./types/*"]
     },
     "noImplicitAny": false,
     "noUnusedParameters": false,
