+++ tsconfig.node.json
@@ -16,7 +16,10 @@
     "strict": true,
     "noUnusedLocals": false,
     "noUnusedParameters": false,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+    "paths": {
+      "@types/*": ["./types/*"]
+    }
   },
   "include": ["vite.config.ts", "types"]
 }
